feat(TDQ-15285): correct initialization-on-demand holder, the init of tokenizer should inside constructor
