+++ dataquality-text-japanese/src/main/java/org/talend/dataquality/jp/tokenization/TextTokenizer.java
@@ -12,25 +12,25 @@
 // ============================================================================
 package org.talend.dataquality.jp.tokenization;
 
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import com.atilika.kuromoji.TokenBase;
-import com.atilika.kuromoji.TokenizerBase;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.talend.dataquality.jp.common.KuromojiDict;
 
-public class TextTokenizer {
+import com.atilika.kuromoji.TokenizerBase;
+
+public class TextTokenizer extends TextTokenizerBase {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(TextTokenizer.class);
 
-    private static TokenizerBase tokenizer;
-
-    private static String dictName;
-
     private TextTokenizer() {
+        // TextTokenizer with dictionary IPADIC
+        final String dictName = KuromojiDict.IPADIC.getDictName();
+        LOGGER.info("Initialise tokenizer with the dictionary: mecab-" + dictName);
+        try {
+            tokenizer = (TokenizerBase) Class.forName("com.atilika.kuromoji." + dictName + ".Tokenizer").newInstance();
+        } catch (Exception e) {
+            LOGGER.error(e.getMessage(), e);
+        }
     }
 
     private static class LazyHolder {
@@ -39,61 +39,7 @@ public class TextTokenizer {
     }
 
     public static TextTokenizer getInstance() throws ClassNotFoundException, InstantiationException, IllegalAccessException {
-        // set default dictionary IPADIC
-        return getInstance(KuromojiDict.IPADIC);
-    }
-
-    public static TextTokenizer getInstance(KuromojiDict dict)
-            throws ClassNotFoundException, InstantiationException, IllegalAccessException {
-        final String dictName = dict.getDictName();
-        if (tokenizer == null || dictName != TextTokenizer.dictName) {
-            LOGGER.info("Initialise tokenizer with the dictionary: mecab-" + dictName);
-            tokenizer = (TokenizerBase) Class.forName("com.atilika.kuromoji." + dictName + ".Tokenizer").newInstance();
-            TextTokenizer.dictName = dictName;
-        }
         return LazyHolder.INSTANCE;
-    }
-
-    /**
-     *
-     * @param text
-     * @return List of TokenBase
-     */
-    public List<? extends TokenBase> tokenize(String text) {
-        return tokenizer.tokenize(text);
-    }
-
-    private Stream<String> getTokenSurface(String text) {
-
-        return this.tokenize(text).stream().map(token -> token.getSurface());
-    }
-
-    /**
-     *
-     * @param text
-     * @return List of tokens
-     */
-    public List<String> getListTokens(String text) {
-        return getTokenSurface(text).collect(Collectors.toList());
-    }
-
-    /**
-     *
-     * @param text
-     * @param delimiter
-     * @return tokenized string with delimiter
-     */
-    public String getTokenizedString(String text, String delimiter) {
-        return getTokenSurface(text).collect(Collectors.joining(delimiter));
-    }
-
-    /**
-     *
-     * @param text
-     * @return tokenized string with space as delimiter
-     */
-    public String getTokenizedString(String text) {
-        return getTokenizedString(text, " ");
     }
 
 }
